refactor(app): split monolithic lib.rs/main.rs into per-module files
Reorganize aionui-app into a layered layout where lib.rs / main.rs / mod.rs
are pure façades (declarations + re-exports only) and every type/function
lives in its own focused file:

- lib.rs: re-exports only — moves AppConfig + derive_encryption_key to
  config.rs, AppServices to services.rs, ModuleStates + build_*_state to
  router/state.rs (merged from former module_states.rs + state_builders.rs).
- router/: routes.rs (create_router*), state.rs (states + builders),
  trace.rs (access log), health.rs (/health + /api/system/guide-mcp).
- main.rs: 47 lines — only CLI parse + runtime init + subcommand dispatch.
- bootstrap/: binary-private startup helpers (tracing_init, environment,
  work_dir, builtin_skills) — each is its own file.
- commands/: only the 4 real subcommands (bridge, server, team_guide,
  team_stdio); previous mixed-in helpers moved out to bootstrap/.

Dependency direction is strictly outer→inner. AppServices stays in the
lib tree because tests/* consume it directly; the side-effect-laden
bootstrap helpers stay in the binary tree (single consumer = main).

No behavioral change. tests pass, clippy clean, fmt clean.
