+++ crates/aionui-app/src/bootstrap/tracing_init.rs
@@ -1,0 +1,95 @@
+//! Tracing subscriber + log file initialization for the binary.
+//!
+//! Lives in the binary tree (not lib) because it owns process-global
+//! subscriber registration that should never be invoked from tests or
+//! external consumers of the library.
+
+use std::path::Path;
+
+use tracing_subscriber::{EnvFilter, Layer, fmt, layer::SubscriberExt, util::SubscriberInitExt};
+
+const NOISE_SUPPRESSIONS: &[&str] = &["sqlx::query=warn", "hyper_util=warn", "reqwest=warn"];
+
+const AIONRS_TARGETS: &[&str] = &[
+    "aion_agent",
+    "aion_config",
+    "aion_compact",
+    "aion_mcp",
+    "aion_providers",
+    "aion_protocol",
+    "aion_tools",
+    "aion_skills",
+    "aion_memory",
+];
+
+fn build_env_filter(log_level: Option<&str>) -> EnvFilter {
+    let user_directives = log_level.unwrap_or("info");
+    let suppressions = NOISE_SUPPRESSIONS.join(",");
+    EnvFilter::new(format!("{suppressions},{user_directives}"))
+}
+
+fn build_backend_filter(log_level: Option<&str>) -> EnvFilter {
+    let user_directives = log_level.unwrap_or("info");
+    let suppressions = NOISE_SUPPRESSIONS.join(",");
+    let aionrs_off: String = AIONRS_TARGETS
+        .iter()
+        .map(|t| format!("{t}=off"))
+        .collect::<Vec<_>>()
+        .join(",");
+    EnvFilter::new(format!("{suppressions},{aionrs_off},{user_directives}"))
+}
+
+/// RAII guards that flush log buffers on drop. Hold for the process lifetime.
+pub struct LogGuards {
+    _backend: tracing_appender::non_blocking::WorkerGuard,
+    _aionrs: tracing_appender::non_blocking::WorkerGuard,
+}
+
+pub fn init_tracing(log_dir: &Path, log_level: Option<&str>) -> LogGuards {
+    std::fs::create_dir_all(log_dir).expect("failed to create log directory");
+
+    let console_layer = fmt::layer().with_target(true).with_filter(build_env_filter(log_level));
+
+    // Backend file layer — excludes aion_* targets
+    let file_appender = tracing_appender::rolling::RollingFileAppender::builder()
+        .rotation(tracing_appender::rolling::Rotation::DAILY)
+        .filename_suffix("backend.log")
+        .build(log_dir)
+        .expect("failed to create backend log file appender");
+    let (non_blocking, backend_guard) = tracing_appender::non_blocking(file_appender);
+
+    let backend_file_layer = fmt::layer()
+        .json()
+        .with_writer(non_blocking)
+        .with_ansi(false)
+        .with_target(true)
+        .with_filter(build_backend_filter(log_level));
+
+    // Aionrs file layer — only aion_* targets
+    let aionrs_level = {
+        let level = log_level.unwrap_or("info");
+        AIONRS_TARGETS
+            .iter()
+            .map(|t| format!("{t}={level}"))
+            .collect::<Vec<_>>()
+            .join(",")
+    };
+    let aionrs_resolved = aion_config::logging::ResolvedLogging {
+        enabled: true,
+        level: aionrs_level,
+        dir: log_dir.to_path_buf(),
+    };
+    let (aionrs_layer, aionrs_guard) =
+        aion_config::logging::create_file_layer(&aionrs_resolved).expect("failed to create aionrs log layer");
+
+    tracing_subscriber::registry()
+        .with(console_layer)
+        .with(backend_file_layer)
+        .with(aionrs_layer)
+        .init();
+
+    LogGuards {
+        _backend: backend_guard,
+        _aionrs: aionrs_guard,
+    }
+}
