+++ crates/aionui-app/src/router/routes.rs
@@ -1,0 +1,256 @@
+//! Top-level router assembly: middleware stack + module route merges.
+
+use std::sync::Arc;
+
+use axum::extract::DefaultBodyLimit;
+use axum::http::Method;
+use axum::middleware::from_fn_with_state;
+use axum::routing::get;
+use axum::{Router, middleware};
+use tower_http::cors::{Any, CorsLayer};
+
+use aionui_ai_agent::{agent_routes, remote_agent_routes};
+use aionui_assets::{AssetRouterState, asset_routes};
+use aionui_assistant::assistant_routes;
+use aionui_auth::{
+    AuthRouterState, AuthState, auth_middleware, auth_routes, csrf_middleware, security_headers_middleware,
+};
+use aionui_channel::channel_routes;
+#[cfg(feature = "weixin")]
+use aionui_channel::weixin_login_route;
+use aionui_conversation::{conversation_ops_routes, conversation_routes};
+use aionui_cron::cron_routes;
+use aionui_extension::{extension_routes, hub_routes, skill_routes};
+use aionui_file::file_routes;
+use aionui_mcp::mcp_routes;
+use aionui_office::{office_proxy_routes, office_routes};
+use aionui_realtime::{WsHandlerState, ws_upgrade_handler};
+use aionui_shell::shell_routes;
+use aionui_system::{connection_test_routes, system_routes};
+use aionui_team::team_routes;
+
+use crate::services::AppServices;
+
+use super::health::{guide_mcp_status, health_check};
+use super::state::{ModuleStates, build_module_states, build_ws_state};
+use super::trace::with_access_log;
+
+/// Create the application router with all routes and global middleware.
+///
+/// Middleware stack (outermost → innermost):
+/// 1. Security response headers (X-Frame-Options, etc.)
+/// 2. CSRF protection (Double Submit Cookie)
+/// 3. Route handlers (auth routes + system routes + conversation routes + file routes + health check)
+pub async fn create_router(services: &AppServices) -> Router {
+    // Bridge event bus → WebSocket manager: forward all broadcast events
+    // to connected WebSocket clients.
+    let mut event_rx = services.event_bus.subscribe();
+    let ws_manager = services.ws_manager.clone();
+    tokio::spawn(async move {
+        while let Ok(event) = event_rx.recv().await {
+            ws_manager.broadcast_all(event);
+        }
+    });
+
+    let (states, channel_components) = build_module_states(services).await;
+
+    // Wire TeamSessionService into Guide MCP server now that both are available.
+    services
+        .inject_guide_service(Arc::downgrade(&states.team.service))
+        .await;
+
+    // Start channel orchestrator (message loop)
+    tokio::spawn(
+        channel_components
+            .orchestrator
+            .run(channel_components.message_rx, channel_components.confirm_rx),
+    );
+
+    // Restore enabled channel plugins (starts receiving IM messages)
+    let chan_mgr = channel_components.manager;
+    let chan_factory = channel_components.plugin_factory;
+    tokio::spawn(async move {
+        if let Err(e) = chan_mgr.restore_plugins(&chan_factory).await {
+            tracing::warn!(error = %e, "failed to restore channel plugins");
+        }
+    });
+
+    create_router_with_states(services, states)
+}
+
+/// Create the application router with custom module states.
+///
+/// Used for testing when specific service overrides are needed
+/// (e.g. injecting a mock HTTP server URL for version check).
+pub fn create_router_with_states(services: &AppServices, states: ModuleStates) -> Router {
+    let ws_state = build_ws_state(services);
+    create_router_with_all_state(services, states, ws_state)
+}
+
+/// Create the application router with custom module states and WebSocket state.
+///
+/// Full-control variant used by tests that need to override
+/// module services and WebSocket behaviour.
+pub fn create_router_with_all_state(services: &AppServices, states: ModuleStates, ws_state: WsHandlerState) -> Router {
+    let auth_state = AuthRouterState {
+        jwt_service: services.jwt_service.clone(),
+        user_repo: services.user_repo.clone(),
+        cookie_config: services.cookie_config.clone(),
+        qr_token_store: services.qr_token_store.clone(),
+        local: services.local,
+    };
+
+    let auth_mw_state = AuthState {
+        jwt_service: services.jwt_service.clone(),
+        user_repo: services.user_repo.clone(),
+        local: services.local,
+    };
+
+    // System routes protected by auth middleware
+    let system_authenticated =
+        system_routes(states.system).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Conversation routes protected by auth middleware
+    let conversation_authenticated = conversation_routes(states.conversation.clone())
+        .route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    let conversation_ops_authenticated = conversation_ops_routes(states.conversation)
+        .route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Remote agent routes protected by auth middleware
+    let remote_agent_authenticated = remote_agent_routes(states.remote_agent)
+        .route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Unified agent listing/refresh/test routes protected by auth middleware
+    let agent_authenticated =
+        agent_routes(states.agent).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Connection test routes (Bedrock, Gemini) protected by auth middleware
+    let connection_test_authenticated = connection_test_routes(states.connection_test)
+        .route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // File routes protected by auth middleware
+    let file_authenticated =
+        file_routes(states.file).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // MCP routes protected by auth middleware
+    let mcp_authenticated =
+        mcp_routes(states.mcp).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Extension routes protected by auth middleware
+    let extension_authenticated =
+        extension_routes(states.extension).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Hub routes protected by auth middleware
+    let hub_authenticated =
+        hub_routes(states.hub).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Skill routes protected by auth middleware
+    let skill_authenticated =
+        skill_routes(states.skill).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Channel routes protected by auth middleware
+    #[cfg(feature = "weixin")]
+    let weixin_login_authenticated = weixin_login_route(states.channel.clone())
+        .route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+    let channel_authenticated =
+        channel_routes(states.channel).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Team routes protected by auth middleware
+    let team_authenticated =
+        team_routes(states.team).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Cron routes protected by auth middleware
+    let cron_authenticated =
+        cron_routes(states.cron).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Office routes protected by auth middleware
+    let office_authenticated =
+        office_routes(states.office.clone()).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Shell + STT routes protected by auth middleware
+    let shell_authenticated =
+        shell_routes(states.shell).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Assistant routes protected by auth middleware (T1a skeleton: all
+    // handlers return 500 "not implemented"; T1b wires real service)
+    let assistant_authenticated =
+        assistant_routes(states.assistant).route_layer(from_fn_with_state(auth_mw_state.clone(), auth_middleware));
+
+    // Guide MCP diagnostic endpoint protected by auth middleware
+    let guide_mcp_authenticated = Router::new()
+        .route("/api/system/guide-mcp", get(guide_mcp_status))
+        .with_state(services.guide_mcp_config.clone())
+        .route_layer(from_fn_with_state(auth_mw_state, auth_middleware));
+
+    // Office proxy routes — exempt from auth (serve iframe content)
+    let office_proxy = office_proxy_routes(states.office);
+    let public_assets = asset_routes(AssetRouterState::default());
+
+    // WebSocket upgrade route — exempt from CSRF (no cookie-based
+    // double-submit) but still gets security response headers.
+    let ws_routes = Router::new().route("/ws", get(ws_upgrade_handler)).with_state(ws_state);
+
+    let router = Router::new()
+        .route("/health", get(health_check))
+        .merge(auth_routes(auth_state))
+        .merge(system_authenticated)
+        .merge(conversation_authenticated)
+        .merge(conversation_ops_authenticated)
+        .merge(remote_agent_authenticated)
+        .merge(agent_authenticated)
+        .merge(connection_test_authenticated)
+        .merge(file_authenticated)
+        .merge(mcp_authenticated)
+        .merge(extension_authenticated)
+        .merge(hub_authenticated)
+        .merge(skill_authenticated)
+        .merge(channel_authenticated)
+        .merge(team_authenticated)
+        .merge(cron_authenticated)
+        .merge(office_authenticated)
+        .merge(shell_authenticated)
+        .merge(assistant_authenticated)
+        .merge(guide_mcp_authenticated);
+
+    // Conditionally merge WeChat login SSE route (feature-gated)
+    #[cfg(feature = "weixin")]
+    let router = router.merge(weixin_login_authenticated);
+
+    let router = if services.local {
+        router
+    } else {
+        router.layer(middleware::from_fn_with_state(
+            services.cookie_config.clone(),
+            csrf_middleware,
+        ))
+    }
+    .merge(ws_routes)
+    .merge(office_proxy)
+    .merge(public_assets)
+    .layer(middleware::from_fn(security_headers_middleware));
+
+    // Raise the default request body limit from axum's 2MB default to
+    // `BODY_LIMIT` (10MB). Routes that need a larger cap (e.g. `/api/fs/upload`)
+    // disable this default and install their own `RequestBodyLimitLayer`.
+    let router = router.layer(DefaultBodyLimit::max(aionui_common::constants::BODY_LIMIT));
+
+    let router = with_access_log(router);
+
+    if services.local {
+        let cors = CorsLayer::new()
+            .allow_origin(Any)
+            .allow_methods([
+                Method::GET,
+                Method::POST,
+                Method::PUT,
+                Method::PATCH,
+                Method::DELETE,
+                Method::OPTIONS,
+            ])
+            .allow_headers(Any);
+        router.layer(cors)
+    } else {
+        router
+    }
+}
