+++ crates/aionui-app/src/services.rs
@@ -1,0 +1,251 @@
+//! Shared application services for dependency injection.
+
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use aionui_ai_agent::{
+    AcpSessionSyncService, AcpSkillManager, AgentFactoryDeps, AgentRegistry, IWorkerTaskManager, WorkerTaskManagerImpl,
+    build_agent_factory,
+};
+use aionui_api_types::GuideMcpConfig;
+use aionui_auth::{CookieConfig, JwtService, QrTokenStore, resolve_jwt_secret};
+use aionui_common::OnConversationDelete;
+use aionui_db::{
+    Database, IAcpSessionRepository, IAgentMetadataRepository, IConversationRepository, IUserRepository,
+    SqliteAcpSessionRepository, SqliteAgentMetadataRepository, SqliteConversationRepository, SqliteProviderRepository,
+    SqliteRemoteAgentRepository, SqliteUserRepository,
+};
+use aionui_realtime::{BroadcastEventBus, WebSocketManager};
+use aionui_team::GuideMcpServer;
+
+use crate::config::{AppConfig, derive_encryption_key};
+
+pub struct AppServices {
+    pub database: Database,
+    pub jwt_service: Arc<JwtService>,
+    pub user_repo: Arc<dyn IUserRepository>,
+    pub cookie_config: Arc<CookieConfig>,
+    pub qr_token_store: Arc<QrTokenStore>,
+    pub ws_manager: Arc<WebSocketManager>,
+    pub event_bus: Arc<BroadcastEventBus>,
+    pub worker_task_manager: Arc<dyn IWorkerTaskManager>,
+    /// Same instance as `worker_task_manager`, exposed through the
+    /// `OnConversationDelete` trait so `ConversationService::with_delete_hook`
+    /// can wire it up. Optional because tests construct `AppServices` with a
+    /// mock `worker_task_manager` that does not implement the trait.
+    pub task_manager_delete_hook: Option<Arc<dyn OnConversationDelete>>,
+    pub agent_registry: Arc<AgentRegistry>,
+    pub conversation_repo: Arc<dyn IConversationRepository>,
+    pub acp_session_sync: Arc<AcpSessionSyncService>,
+    /// Raw JWT secret string, used to derive encryption keys.
+    pub jwt_secret_raw: String,
+    pub data_dir: PathBuf,
+    pub work_dir: PathBuf,
+    /// When `true`, skip JWT authentication and use a fixed default user.
+    pub local: bool,
+    pub app_version: String,
+    /// Resolved skill paths. Shared with the `ConversationService` for
+    /// snapshot resolution at create time.
+    pub skill_paths: Arc<aionui_extension::SkillPaths>,
+    /// Guide MCP server config (port, token, binary_path).
+    /// `None` when the server failed to start (graceful degradation).
+    pub guide_mcp_config: Option<GuideMcpConfig>,
+    /// Guide MCP server instance kept alive for the app lifetime.
+    pub(crate) _guide_server: Option<GuideMcpServer>,
+}
+
+impl AppServices {
+    /// Replace the worker task manager after construction.
+    ///
+    /// Primarily used by tests to inject mock implementations.
+    pub fn with_worker_task_manager(mut self, wtm: Arc<dyn IWorkerTaskManager>) -> Self {
+        self.worker_task_manager = wtm;
+        self
+    }
+
+    /// Wire the TeamSessionService into the Guide MCP server so
+    /// `aion_create_team` requests can call `service.create_team(...)`.
+    /// Called from `create_router` after `build_module_states`.
+    pub(crate) async fn inject_guide_service(&self, service: std::sync::Weak<aionui_team::TeamSessionService>) {
+        if let Some(server) = &self._guide_server {
+            server.set_service(service).await;
+        }
+    }
+
+    pub async fn from_config(database: Database, config: &AppConfig) -> anyhow::Result<Self> {
+        let data_dir = config.data_dir.clone();
+        let work_dir = config.work_dir.clone();
+        let local = config.local;
+        let app_version = config.app_version.clone();
+        let user_repo: Arc<dyn IUserRepository> = Arc::new(SqliteUserRepository::new(database.pool().clone()));
+
+        // Resolve JWT secret: env var → system user db field → random generation
+        let env_secret = std::env::var("JWT_SECRET").ok();
+        let system_user = user_repo
+            .get_system_user()
+            .await
+            .map_err(|e| anyhow::anyhow!("Failed to get system user: {e}"))?;
+
+        let db_secret = system_user
+            .as_ref()
+            .and_then(|u| u.jwt_secret.as_deref())
+            .filter(|s| !s.is_empty());
+
+        let (secret, is_new) = resolve_jwt_secret(env_secret.as_deref(), db_secret);
+
+        // Persist newly generated secret to database
+        if is_new && let Some(user) = &system_user {
+            user_repo
+                .update_jwt_secret(&user.id, &secret)
+                .await
+                .map_err(|e| anyhow::anyhow!("Failed to persist JWT secret: {e}"))?;
+            tracing::info!("Generated and persisted new JWT secret");
+        }
+
+        let encryption_key = derive_encryption_key(&secret);
+
+        let remote_agent_repo = Arc::new(SqliteRemoteAgentRepository::new(database.pool().clone()));
+        let provider_repo = Arc::new(SqliteProviderRepository::new(database.pool().clone()));
+
+        let agent_metadata_repo: Arc<dyn IAgentMetadataRepository> =
+            Arc::new(SqliteAgentMetadataRepository::new(database.pool().clone()));
+        let agent_registry = AgentRegistry::new(agent_metadata_repo);
+        agent_registry
+            .hydrate()
+            .await
+            .map_err(|e| anyhow::anyhow!("Failed to hydrate agent registry: {e}"))?;
+
+        let acp_session_repo: Arc<dyn IAcpSessionRepository> =
+            Arc::new(SqliteAcpSessionRepository::new(database.pool().clone()));
+        let acp_agent_service = AcpSessionSyncService::new(acp_session_repo.clone());
+
+        let conversation_repo: Arc<dyn IConversationRepository> =
+            Arc::new(SqliteConversationRepository::new(database.pool().clone()));
+
+        // Skill paths need app resource dir (for builtin rules) + data dir
+        // (for user skills + materialized views). AcpSkillManager uses these
+        // for first-message skill index/body loading.
+        let app_resource_dir = std::env::current_exe()
+            .ok()
+            .and_then(|p| p.canonicalize().ok())
+            .and_then(|p| p.parent().map(|pp| pp.to_path_buf()))
+            .unwrap_or_else(|| std::path::PathBuf::from("."));
+        let skill_paths = Arc::new(aionui_extension::resolve_skill_paths(&app_resource_dir, &data_dir));
+
+        // Absolute path to this process's binary. Reused as the `command` for
+        // the stdio MCP bridge spawned by ACP CLIs when a team session is
+        // attached to a conversation (phase1 mcp.md §4.6 single-binary model).
+        let backend_binary_path =
+            Arc::new(std::env::current_exe().unwrap_or_else(|_| std::path::PathBuf::from("aionui-backend")));
+
+        // Start Guide MCP server. Failure is non-fatal: solo agents simply
+        // won't get the `aion_create_team` tool.
+        let (guide_server, guide_mcp_config) = match GuideMcpServer::start().await {
+            Ok(srv) => {
+                let config = GuideMcpConfig {
+                    port: srv.http_port(),
+                    token: srv.auth_token().to_owned(),
+                    binary_path: backend_binary_path.to_string_lossy().to_string(),
+                };
+                tracing::info!(port = config.port, "Guide MCP server started");
+                (Some(srv), Some(config))
+            }
+            Err(e) => {
+                tracing::warn!(error = %e, "Guide MCP server failed to start; solo create-team disabled");
+                (None, None)
+            }
+        };
+
+        let factory = build_agent_factory(AgentFactoryDeps {
+            skill_manager: AcpSkillManager::new(skill_paths.clone()),
+            remote_agent_repo,
+            provider_repo,
+            encryption_key,
+            agent_registry: agent_registry.clone(),
+            acp_agent_service: acp_agent_service.clone(),
+            data_dir: data_dir.clone(),
+            backend_binary_path: backend_binary_path.clone(),
+            guide_mcp_config: guide_mcp_config.clone(),
+        });
+
+        // Agent factory is now wired. Future extension/custom agents
+        // that get written to `agent_metadata` will show up after the
+        // relevant service calls `AgentRegistry::hydrate`.
+        let task_manager_concrete = Arc::new(WorkerTaskManagerImpl::new(factory));
+        let worker_task_manager: Arc<dyn IWorkerTaskManager> = task_manager_concrete.clone();
+        let task_manager_delete_hook: Arc<dyn OnConversationDelete> = task_manager_concrete;
+
+        Ok(Self {
+            database,
+            jwt_service: Arc::new(JwtService::new(secret.clone())),
+            user_repo,
+            cookie_config: Arc::new(CookieConfig::from_env()),
+            qr_token_store: Arc::new(QrTokenStore::new()),
+            ws_manager: Arc::new(WebSocketManager::new()),
+            event_bus: Arc::new(BroadcastEventBus::new(256)),
+            worker_task_manager,
+            task_manager_delete_hook: Some(task_manager_delete_hook),
+            agent_registry,
+            conversation_repo,
+            acp_session_sync: acp_agent_service,
+            jwt_secret_raw: secret,
+            data_dir,
+            work_dir,
+            local,
+            app_version,
+            skill_paths,
+            guide_mcp_config: guide_mcp_config.clone(),
+            _guide_server: guide_server,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_app_services_from_memory_db() {
+        let db = aionui_db::init_database_memory().await.unwrap();
+        let services = AppServices::from_config(db, &AppConfig::default()).await.unwrap();
+
+        // JWT service should be functional
+        let token = services.jwt_service.sign("test_user", "testuser").unwrap();
+        let payload = services.jwt_service.verify(&token).unwrap();
+        assert_eq!(payload.user_id, "test_user");
+
+        // User repo should have system user
+        let has_users = services.user_repo.has_users().await.unwrap();
+        assert!(!has_users); // system user has empty password → not counted
+
+        services.database.close().await;
+    }
+
+    #[tokio::test]
+    async fn test_jwt_secret_persisted_to_db() {
+        let db = aionui_db::init_database_memory().await.unwrap();
+        let services = AppServices::from_config(db, &AppConfig::default()).await.unwrap();
+
+        // System user should now have a jwt_secret persisted
+        let system_user = services.user_repo.get_system_user().await.unwrap();
+        let jwt_secret = system_user.unwrap().jwt_secret;
+        assert!(jwt_secret.is_some());
+        assert!(!jwt_secret.unwrap().is_empty());
+
+        services.database.close().await;
+    }
+
+    #[tokio::test]
+    async fn test_app_services_uses_supplied_app_version() {
+        let db = aionui_db::init_database_memory().await.unwrap();
+        let config = AppConfig {
+            app_version: "9.9.9".to_string(),
+            ..Default::default()
+        };
+        let services = AppServices::from_config(db, &config).await.unwrap();
+
+        assert_eq!(services.app_version, "9.9.9");
+
+        services.database.close().await;
+    }
+}
